// packages/generator/src/adapters/SerializationAdapter.d.ts
export interface MessageDescriptor {
    /**
     * Fully qualified message name (e.g., "test.services.GetUserRequest")
     */
    messageName: string;
    /**
     * Field descriptors for this message
     */
    fields: FieldDescriptor[];
}
/**
 * Field descriptor metadata
 */
export interface FieldDescriptor {
    /**
     * Field name in camelCase
     */
    name: string;
    /**
     * Field number from proto definition
     */
    number: number;
    /**
     * Field type (e.g., "string", "int32", "bytes", "Message")
     */
    type: string;
    /**
     * Whether this field is repeated (array)
     */
    repeated: boolean;
    /**
     * Whether this field is optional
     */
    optional: boolean;
    /**
     * Whether this field is a map
     */
    map?: boolean;
    /**
     * For message types, the message descriptor
     */
    messageType?: MessageDescriptor;
}
// ...
export declare class SerializationError extends Error {
    readonly field?: string | undefined;
    readonly value?: any | undefined;
    constructor(message: string, field?: string | undefined, value?: any | undefined);
}